move backfill seer utils to separate module
